refactor(profile): split ProfileContent into read-only and editable variants
ProfileContent in publicprofile/ is now a pure read-only display with
named slots at each editable insertion point. The previous edit-aware
version moves to myprofile/EditableProfileContent.vue as a thin wrapper
that fills those slots with EditField instances.

Public profile and conversation panel views no longer pull EditField,
useEditFields, or the form editor components (PublicNameInput,
LocationSelector, LanguageSelector, TagSelector, IntrotextEditor,
ImageEditor) into their bundles. MyProfile swaps to the editable wrapper
and drops a couple of dead props/listeners on the way through.
